fix: clear cross-batch review lint blockers
